+++ compiler/src/iree/compiler/Dialect/Util/IR/BUILD.bazel
@@ -52,7 +52,6 @@ iree_compiler_cc_library(
     ],
     hdrs = [
         "ClosureOpUtils.h",
-        "IntegerDivisibility.h",
         "UtilDialect.h",
         "UtilOps.h",
         "UtilTraits.h",
